fix: user retrieve and update using JWT instead of getting the username from endpoint

+++ authentication/views.py
@@ -94,7 +94,9 @@ class UserRetriveView(APIView):
     serializer_class= UserRetriveSerializer
     
     def get(self, request, *args, **kwargs):
-        username = kwargs.get('username')
+        if (not request.user.is_authenticated):
+            return Response("user not authenticated", status=status.HTTP_401_UNAUTHORIZED)
+        username = request.user.username
         user = User.objects.get(username=username)
         serializer = self.serializer_class(user)
         return Response(serializer.data, status=status.HTTP_200_OK)
@@ -104,20 +106,21 @@ class UserDeleteView(APIView):
     serializer_class= UserDeleteSerializer
     
     def post(self, request, *args, **kwargs):
-        try:
-            username = kwargs.get('username')
-            user = User.objects.get(username=username)
-            serializer = self.serializer_class(data = request.data)
-            
-            serializer.is_valid(raise_exception=True)
-            
-            if user.check_password(serializer.validated_data.get('password')):
-                return Response("invalid password", status=status.HTTP_400_BAD_REQUEST)
-            
-            user.delete()
-            return Response("user deleted successfully", status=status.HTTP_200_OK)
-        except User.DoesNotExist:
-            return Response("user not found", status=status.HTTP_404_NOT_FOUND)
+        if (not request.user.is_authenticated):
+            return Response("user not authenticated", status=status.HTTP_401_UNAUTHORIZED)
+        
+        username = request.user.username
+        user = User.objects.get(username=username)
+        serializer = self.serializer_class(data = request.data)
+        
+        serializer.is_valid(raise_exception=True)
+        
+        if user.check_password(serializer.validated_data.get('password')):
+            return Response("invalid password", status=status.HTTP_400_BAD_REQUEST)
+        
+        user.delete()
+        return Response("user deleted successfully", status=status.HTTP_200_OK)
+
         
 
 class UserUpdateView(APIView):
@@ -126,7 +129,7 @@ class UserUpdateView(APIView):
     
     def patch(self, request, *args, **kwargs):
         try:
-            username = kwargs.get('username')
+            username = request.user.username
             user = User.objects.get(username=username)
             serializer = self.serializer_class(user, data = request.data, partial=True)
